Switch base OCI images from Debian to Alpine
Alpine is significantly smaller than even "slim" Debian images, so
development should go quicker if we use them.

Here are the differences in the base image sizes:

- `20.18.0-alpine3.20` is 45.42 MB while `20.18.0-bookworm-slim` is 68.28 MB
- `3.7.17-alpine3.18` is 16.99 MB while `3.7.17-slim-bookworm` is 44.95 MB

This drops the size of the final images from 80.7 MB to 38.9 MB. It's not
much, but it should still help some. It has the added benefit of having
fewer vulnerabilities, which is important because the Python 3.7 base
images haven't been updated in over a year.

Changing the base OS required multiple related build changes:

- Update packages with `apk` instead of `apt-get`
- Run server with `sh` because `bash` is not installed by default

I am only switching this repository over to using Alpine base images
because its dependencies are simple. I'm not going to attempt running
CUDA or any other complex dependency on such a basic image.